// Client/StockWebFrontEnd/src/types/tableDataType.tsx
export interface tableDataType {
    key: React.Key;
    id: number;
    Date: string;
    Item: string;
    T1_1: number;
    T1_2: number;
    T1_3: number;
    T2_1: number;
    T2_2: number;
    T2_3: number;
    T3_1: number;
    T3_2: number;
    T3_3: number;
    T4_1: number;
    T4_2: number;
    T4_3: number;
    Type: string;
    Delta: number;
    Time: string;
    ValVN30: number;
    VolMax5m: number;
}